Small fix for MetaDataHandler
git-svn-id: https://svn.tribler.org/abc/branches/release-5.5.x@25014 001aeff7-3401-0410-a489-f7902fc005dd

+++ Tribler/Core/Overlay/MetadataHandler.py
@@ -317,7 +317,14 @@ class MetadataHandler:
 
 
             return self.do_send_metadata(permid, torrent, selversion)
-        else:    # deleted before sending it
+        
+        else:
+            file_name = get_collected_torrent_filename(infohash)
+            torrent_path2 = os.path.join(self.torrent_dir, file_name)
+            if os.path.exists(torrent_path2) and not os.path.samefile(torrent_path, torrent_path2):
+                return self.read_and_send_metadata(permid, infohash, torrent_path2, selversion)
+            
+            # deleted before sending it
             self.torrent_db.deleteTorrent(infohash, delete_file=True, commit=True)
             if DEBUG:
                 print >> sys.stderr,"metadata: GET_METADATA: no torrent data to send"
